Fix static analysis
- Remove PHPStan workarounds and require fixed PHPCS version
- PHAN doesn't hard depend on AST anymore in recent versions

+++ MO4/Sniffs/Commenting/PropertyCommentSniff.php
@@ -160,7 +160,7 @@ class PropertyCommentSniff extends AbstractScopeSniff
             $isCommentOneLiner
                 = $tokens[$commentStart]['line'] === $tokens[$commentEnd]['line'];
 
-            $length         = (int) ($commentEnd - $commentStart + 1);
+            $length         = ($commentEnd - $commentStart + 1);
             $tokensAsString = $phpcsFile->getTokensAsString(
                 $commentStart,
                 $length
